Drop tflite and flatbuffer dependencies
There is not real need for the linux_driver_stack to know about tflite
and flatbuffers.  A better approach is to just pass the buffer to the
firmware to be processes, i.e., either parsed or executed.

This solves issues when linux_driver_stack and firmware are not in
sync with the same library versions.

Change-Id: I9b2a12e69f37f61b1ac594433a15883fb1c67b9c